Fix incompatible pointers in audio_read.c

+++ audio_read.c
@@ -16,7 +16,7 @@
 *
 ************************************************************************/
 
-unsigned long read_samples (FILE * musicin, long int *sample_buffer,
+unsigned long read_samples (FILE * musicin, int32_t *sample_buffer,
 			    long unsigned int num_samples,
 			    long unsigned int frame_size, int
 			    *byte_per_sample, int *aiff)
